requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.1
djangorestframework==3.16.1
idna==3.11
kombu==5.6.2
# mysqlclient @ file:///D:/bld/bld/rattler-build_mysqlclient_1767654208/work
# packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049137378/work
mysqlclient
packaging
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
setuptools==80.10.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
wheel==0.46.3
